requirements.txt
dj-database-url==1.3.0
Django==4.1.7
django-cors-headers==3.14.0
djangorestframework==3.14.0
gunicorn==20.1.0
psycopg2-binary==2.9.6
python-dotenv==1.0.0
sqlparse==0.4.3
whitenoise==6.4.0